Revert output mgr stuff (#2063)
* Revert "Fix issue with global output manager tree (#2062)"

This reverts commit 2346aa90663c10e2c2125efac09651be44160c57.

* Revert "Remove all output mgmt from the resolver (#2060)"

This reverts commit f1140b1d3a9acbc6fa165af9f2c49c7d7b8253b4.
