+++ modal/_resolver.py
@@ -1,5 +1,6 @@
 # Copyright Modal Labs 2023
 import asyncio
+import contextlib
 from asyncio import Future
 from typing import TYPE_CHECKING, Dict, Hashable, List, Optional
 
@@ -11,7 +12,33 @@ from .config import logger
 from .exception import NotFoundError
 
 if TYPE_CHECKING:
+    from rich.tree import Tree
+
     from modal.object import _Object
+
+
+class StatusRow:
+    def __init__(self, progress: "Optional[Tree]"):
+        from ._output import (
+            step_progress,
+        )
+
+        self._spinner = None
+        self._step_node = None
+        if progress is not None:
+            self._spinner = step_progress()
+            self._step_node = progress.add(self._spinner)
+
+    def message(self, message):
+        if self._spinner is not None:
+            self._spinner.update(text=message)
+
+    def finish(self, message):
+        from ._output import substep_completed
+
+        if self._step_node is not None:
+            self._spinner.update(text=message)
+            self._step_node.label = substep_completed(message)
 
 
 class Resolver:
@@ -28,7 +55,12 @@ class Resolver:
         environment_name: Optional[str] = None,
         app_id: Optional[str] = None,
     ):
+        from rich.tree import Tree
+
+        from ._output import step_progress
+
         self._local_uuid_to_future = {}
+        self._tree = Tree(step_progress("Creating objects..."), guide_style="gray50")
         self._client = client
         self._app_id = app_id
         self._environment_name = environment_name
@@ -123,3 +155,19 @@ class Resolver:
             obj = fut.result()
             unique_objects.setdefault(obj.object_id, obj)
         return list(unique_objects.values())
+
+    @contextlib.contextmanager
+    def display(self):
+        # TODO(erikbern): get rid of this wrapper
+        from ._output import OutputManager, step_completed
+
+        if output_mgr := OutputManager.get():
+            with output_mgr.make_live(self._tree):
+                yield
+            self._tree.label = step_completed("Created objects.")
+            output_mgr.print(self._tree)
+        else:
+            yield
+
+    def add_status_row(self) -> StatusRow:
+        return StatusRow(self._tree)
